+++ cli/tests/golden/artifacts/rust/customer/customer_customer_event_v1.rs
@@ -84,41 +84,8 @@ impl From<CustomerCustomerV1Status> for CustomerCustomerEventV1Status {
     fn from(src: CustomerCustomerV1Status) -> Self {
         match src {
             CustomerCustomerV1Status::Active => CustomerCustomerEventV1Status::Active,
-            CustomerCustomerV1Status::Deleted => CustomerCustomerEventV1Status::Deleted,
             CustomerCustomerV1Status::Suspended => CustomerCustomerEventV1Status::Suspended,
-        }
-    }
-}
-
-use super::customer_customer_db_v1::CustomerCustomerDbV1Status;
-impl From<CustomerCustomerDbV1Status> for CustomerCustomerEventV1Status {
-    fn from(src: CustomerCustomerDbV1Status) -> Self {
-        match src {
-            CustomerCustomerDbV1Status::Active => CustomerCustomerEventV1Status::Active,
-            CustomerCustomerDbV1Status::Deleted => CustomerCustomerEventV1Status::Deleted,
-            CustomerCustomerDbV1Status::Suspended => CustomerCustomerEventV1Status::Suspended,
-        }
-    }
-}
-
-use super::customer_customer_request_v1::CustomerCustomerRequestV1Status;
-impl From<CustomerCustomerRequestV1Status> for CustomerCustomerEventV1Status {
-    fn from(src: CustomerCustomerRequestV1Status) -> Self {
-        match src {
-            CustomerCustomerRequestV1Status::Active => CustomerCustomerEventV1Status::Active,
-            CustomerCustomerRequestV1Status::Deleted => CustomerCustomerEventV1Status::Deleted,
-            CustomerCustomerRequestV1Status::Suspended => CustomerCustomerEventV1Status::Suspended,
-        }
-    }
-}
-
-use super::customer_customer_reply_v1::CustomerCustomerReplyV1Status;
-impl From<CustomerCustomerReplyV1Status> for CustomerCustomerEventV1Status {
-    fn from(src: CustomerCustomerReplyV1Status) -> Self {
-        match src {
-            CustomerCustomerReplyV1Status::Active => CustomerCustomerEventV1Status::Active,
-            CustomerCustomerReplyV1Status::Deleted => CustomerCustomerEventV1Status::Deleted,
-            CustomerCustomerReplyV1Status::Suspended => CustomerCustomerEventV1Status::Suspended,
+            CustomerCustomerV1Status::Deleted => CustomerCustomerEventV1Status::Deleted,
         }
     }
 }